Correct favicon handling.
As per https://github.com/plone/Products.CMFPlone/pull/3418#issuecomment-1027118882

1. Remove alternate icon HTML, since we do not serve alternate icons,
since we do not actually provide the user with a mechanism to upload
alternate icons, and serving the default Plone icon as an alternate
would cause branding problems for site operators.
2. Hardcode icon URL to /favicon.ico, rooted in the site navigation
root, rather than the overall site root, as the icon serving is done per
site navigation root, not per site root.
3. Ignore file name uploaded by the uploader.  There was no point in
serving the file name, since the view that actually serves the data (in
CMFPlone) cannot "rename itself" according to the file name of the icon
that the user uploaded when he uploaded the custom icon.
4. Instead of informing the browser that the favicon is at /favicon or
/favicon.ico depending on whether the user has customized the favicon,
simply inform the browser that the favicon is at /favicon.ico, and the
CMFPlone view is in charge of serving the correct data to the user.