Preserve the file name cache-busting properties of customizable icons.
When an icon is customized, we can exploit the file name of the customized icon to provide for a cache bust effect -- the cache key (URL) of the favicon URL served to the browser via HTML will vary depending on the name of the uploaded favicon.

We, of course, still preserve full compatibility with direct /favicon.ico access.